♼ Move functions from utils.py to tasks.py
