+++ corehq/motech/dhis2/tasks.py
@@ -1,25 +1,25 @@
-from datetime import datetime
+from datetime import date, datetime, timedelta
 from itertools import chain
 
 from celery.schedules import crontab
 from celery.task import periodic_task, task
+from dateutil.relativedelta import relativedelta
 
+from dimagi.utils.dates import DateSpan
 from toggle.shortcuts import find_domains_with_toggle_enabled
 
 from corehq import toggles
+from corehq.apps.userreports.models import ReportConfiguration
+from corehq.apps.userreports.reports.data_source import (
+    ConfigurableReportDataSource,
+)
 from corehq.motech.dhis2.const import (
     SEND_FREQUENCY_MONTHLY,
     SEND_FREQUENCY_QUARTERLY,
     SEND_FREQUENCY_WEEKLY,
 )
 from corehq.motech.dhis2.dbaccessors import get_dataset_maps
-from corehq.motech.dhis2.utils import (
-    get_date_filter,
-    get_previous_month,
-    get_previous_quarter,
-    get_report_config,
-    get_ucr_data,
-)
+from corehq.util.couch import get_document_or_not_found
 from corehq.util.quickcache import quickcache
 
 
@@ -112,23 +112,6 @@ def get_dataset(dataset_map, send_date):
     return dataset
 
 
-@quickcache(['dataset_map.domain', 'dataset_map.ucr_id'])
-def get_datavalue_map_dict(dataset_map):
-    dict_ = {dvm.column: dict(dvm, is_org_unit=False, is_period=False)
-             for dvm in dataset_map.datavalue_maps}
-    if dataset_map.org_unit_column:
-        dict_[dataset_map.org_unit_column] = {
-            'is_org_unit': True,
-            'is_period': False,
-        }
-    if dataset_map.period_column:
-        dict_[dataset_map.period_column] = {
-            'is_org_unit': False,
-            'is_period': True,
-        }
-    return dict_
-
-
 def get_datavalues(dataset_map, ucr_row):
     """
     Returns rows of "dataElement", "categoryOptionCombo", "value", and
@@ -172,3 +155,78 @@ def get_datavalues(dataset_map, ucr_row):
             if org_unit:
                 datavalue['orgUnit'] = org_unit
     return datavalues
+
+
+@quickcache(['dataset_map.domain', 'dataset_map.ucr_id'])
+def get_datavalue_map_dict(dataset_map):
+    dict_ = {dvm.column: dict(dvm, is_org_unit=False, is_period=False)
+             for dvm in dataset_map.datavalue_maps}
+    if dataset_map.org_unit_column:
+        dict_[dataset_map.org_unit_column] = {
+            'is_org_unit': True,
+            'is_period': False,
+        }
+    if dataset_map.period_column:
+        dict_[dataset_map.period_column] = {
+            'is_org_unit': False,
+            'is_period': True,
+        }
+    return dict_
+
+
+def get_report_config(domain_name, ucr_id):
+    report_config = get_document_or_not_found(ReportConfiguration, domain_name, ucr_id)
+    return report_config
+
+
+def get_date_filter(report_config):
+    """
+    Returns the first date filter, or None.
+
+    Assumes the first date filter is the one to use.
+
+    .. NOTE: The user might not want to filter by date for DHIS2
+             integration. They can use a "period" column to return
+             rows for multiple periods, or set a period for the report
+             if it is always for the same period.
+    """
+    date_filter = next((f for f in report_config.filters if f['type'] == 'date'), None)
+    return date_filter
+
+
+def get_previous_month(send_date: date) -> DateSpan:
+    enddate = date(year=send_date.year, month=send_date.month, day=1) - timedelta(days=1)
+    startdate = date(year=enddate.year, month=enddate.month, day=1)
+    return DateSpan(startdate, enddate)
+
+
+def get_previous_quarter(send_date: date) -> DateSpan:
+    current_quarter_start = (((send_date.month - 1) // 3) * 3) + 1
+    startdate = date(year=send_date.year, month=current_quarter_start, day=1) - relativedelta(months=3)
+    enddate = date(year=send_date.year, month=current_quarter_start, day=1) + relativedelta(months=4) - \
+        timedelta(days=1) - relativedelta(months=3)
+    return DateSpan(startdate, enddate)
+
+
+def get_ucr_data(report_config, date_filter, date_span):
+    from corehq.apps.userreports.reports.view import get_filter_values
+
+    data_source = ConfigurableReportDataSource.from_spec(report_config, include_prefilters=True)
+
+    filter_params = get_date_params(date_filter['slug'], date_span) if date_filter else {}
+    filter_values = get_filter_values(report_config.ui_filters, filter_params)
+    data_source.set_filter_values(filter_values)
+    return data_source.get_data()
+
+
+def get_date_params(slug, date_span):
+    """
+    Mimics date filter request parameters
+    """
+    startdate = date_span.startdate.strftime('%Y-%m-%d')
+    enddate = date_span.enddate.strftime('%Y-%m-%d')
+    return {
+        slug: "{}+to+{}".format(startdate, enddate),
+        slug + '-start': startdate,
+        slug + '-end': enddate,
+    }
